Migrate naive id transformer (#661)
Summary:
This PR is trying to migrate `NaiveIDTransformer` to the top folder. The idea behind the naive transformer is quite simple, we use a hash map (`ska::flat_hash_map`) to store the global id to cache id map, and use a bitmap to record the empty slot in the cache.

cc reyoung

Pull Request resolved: https://github.com/pytorch/torchrec/pull/661

Reviewed By: zyan0

Differential Revision: D39857301

Pulled By: colin2328

fbshipit-source-id: a7681826827d84d51dcf5d91b3a0eca7583f391e