Classify PlannerError based on storage limit (#690)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/690

Some users rely on the PlannerError message to determine whether the "unable to find a plan" error is due to insufficient storage. If so, users should input larger storage instead of trying to tune other inputs (e.g. constraints).

This diff proposes a way to classify the planner error due to "insufficient storage", so users may behave accordingly from the error type.

Differential Revision: D39982352

fbshipit-source-id: 16d5361df10a667eab9443b88c91b6073bd805c0