Early stopping feature (#696)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/696

## Problem
When there are many proposals to plan with (e.g. 1000 tables, each of which is with [fused, fused_uvm_caching] compute_kernels, can result in 1000 proposals in GreedyProposer), the planner runs >20 min to come up with the best plan. The planning step is too long and may fail to broadcast the plan since the collective call waits for 30 min to close the communication.

For GreedyProposer, we found the following trend on the perf_rating (see the graph below): the GreedyProposer firstly proposes a proposal with the lowest perf (all tables put in hbm). If the proposal is not partitionable, GreedyProposer finds the largest table, puts it in uvm, and proposes again. After it reaches a partitionable state, there is a "best_plan" range, i.e., with some tables put in uvm, there is enough space left in hbm for planning, making the perf balanced, resulting in lowest max_perf (perf_rating). After this range is passed, more and more tables are put in UVM, so perf_rating keeps increasing. For MIMO, we observed that it takes ~260 iterations to reach the best_plan range. After several tens of iterations, it reaches the increasing perf_rating range, leaving around ~700 iterations in this range.

{F775152524}

## Solution
To avoid waste of time in the last stage, this diff introduces an "early_stopping" feature to `GreedyProposer` to end the proposing iteration. When early_stopping is turned on (default is false), it observes the passed perf_rating in feedback and updates the best_perf_rating. When the following proposals have worse perf_rating than best_perf_rating, it stops proposing new proposals.
With this feature, the planning time decreases from 20 min to ~5 min. With this feature plus the deepcopy refactor diff (D39822605 (https://github.com/pytorch/torchrec/commit/77db94c88b57040a9f32c09fc805b676ba35b85e)), the planning time can further decreased to <1 min.

Reviewed By: dstaay-fb

Differential Revision: D39068786

fbshipit-source-id: 43fed48c7e73b06233f35fa1e8f9ae2c25133c65