Fix bug in storage comparison
Summary: When using "<" to compare `Storage`, it would only compare the hbm size, regardless of ddr size. This diff changes that to `fits_in` to compare the Storage (lowest_storage vs. global_storage_constraints), so that proper planner error type is raised when either hbm or ddr of lowest_storage is greater than that of global_storage_constraints.

Differential Revision: D40349774

fbshipit-source-id: e2cdd1906c6859570610812caadbbfd0eca265b6