Fix bug in storage comparison (#722)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/722

When using "<" to compare `Storage`, it would only compare the hbm size, regardless of ddr size. This diff changes that to `fits_in` to compare the Storage (lowest_storage vs. global_storage_constraints), so that proper planner error type is raised when either hbm or ddr of lowest_storage is greater than that of global_storage_constraints.

Reviewed By: joshuadeng

Differential Revision: D40349774

fbshipit-source-id: db9f98ab50c689fafabea641c0a9f5e00a23a603