Migrate MixedLFULRUStrategy (#701)
Summary:
This PR is trying to migrate `MixedLFULRUStrategy` to the upper-level folder. The main functionality of this class is to help decide which global_ids to evict when the id transformer (for example, the `NaiveIDTransformer`) is full.

As its name suggested, it's a mix of LFU and LRU. We got the idea for this strategy from [redis](https://redis.io/docs/manual/eviction/#the-new-lfu-mode): we only need an approximate counter for the frequency of each global id stored in the id transformer.

Specifically, we will roll the dice - checking the next k bits of the random generated bits - each time we visit an id, and if all k bits are 0, we will increase the counter. This makes the probability of increment $1/2^k$. And by increasing the k along with the increment of the counter, we can record an approximate frequency counter with few bits, squeeze both the timestamp (the LRU part) and frequency counter (the LFU part) in a uint32_t.

Thank you for your time on this PR :)

cc reyoung colin2328

Pull Request resolved: https://github.com/pytorch/torchrec/pull/701

Reviewed By: zyan0

Differential Revision: D40325576

Pulled By: colin2328

fbshipit-source-id: 450369b305951a73453702de8a4cbfddf823540c