quant ebc
Summary:
-- Motivation --

We may want to script the unsharded QEBC as part of inference (offload it to CPU). However, scripting the original quant EBC is a nightmare as many things aren't supported (dataclasses, for loop etc). scripting the traced module is much easier since everything is static/collapsed.

However, tracing has the below issue:

Since torchrec doesn't want to register in a module list, we need to make it compatible with fx tracing

In the forward, it looks at all modules encountered, but needs a name. Usually this is derived from named_modules(). But since we don't register things as modules, we need to give some help to fx to identify the name.

Reviewed By: YazhiGao

Differential Revision: D40724168

fbshipit-source-id: 0fa93aacd24d29d65334205df9e74c97ad666932