Optimize ETC input dist for new kjt a2a (#854)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/854

Currently the ETC input dist pattern is that it does the cross_pg input dist (inter host) and waits for it to finish before starting the intra host input dist.

With the new KJT a2a updates we have the initial collective call send splits which is smaller metadata information before sending the tensors. Since the inter host comms is more expensive, we want to optimize the input dist by only starting the cross dist/inter host input dist initially. This will only block the cpu for the minimum amount of time in the sparse data dist step because only sending the splits would be synchronous, whereas before we would have to wait for the entire inter host input dist to finish and then start the intra host input dist.

As a result this delays the intra host input dist until the forward step, which will slow it down, however, intra host comms are faster, so holistically the time saved in the sparse data dist makes up for the extra time taken in the forward with the intra host input dist.

Reviewed By: bigning

Differential Revision: D41638406

fbshipit-source-id: 0b7418c1c5476c3ecaf13569ee53df6dcede8c49