Create optimizer prehook objects (#864)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/864

A change was made to torch.optim.Optimizer that adds these initialize states.

However, we don't call super() because optimizer.super calls add_param_group, which is marked as not implemented explicitly.

Adding these to unbreak trunk, but IMO we should rethink how we do add_param_Groups. It's hard to keep updating this main file with all changes in optimizer.py

Differential Revision: D41774123

fbshipit-source-id: 42250899253f5c2d379d2bba108ce9cbe8c5cb41