Create optimizer prehook objects (#864)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/864

A change was made to torch.optim.Optimizer that adds these initialize states.

However, we don't call super() because optimizer.super calls add_param_group, which is marked as not implemented explicitly.

Adding these to unbreak trunk, but IMO we should rethink how we do add_param_Groups. It's hard to keep updating this main file with all changes in optimizer.py

Reviewed By: colin2328

Differential Revision: D41774123

fbshipit-source-id: 6988d55916aaaf82fb9cabc051d3da551f1b764d