Correct confusing naming in _calculate_cw_shard_sizes_and_offsets
Summary: `hash_size` are now used in both `_calculate_rw_shard_sizes_and_offsets` and `_calculate_cw_shard_sizes_and_offsets`, but carry different meanings. It means `rows` in the former (this is correct) but means `columns` in the latter (this is wrong). We should make it correct and consistent.

Reviewed By: joshuadeng

Differential Revision: D42640749

fbshipit-source-id: a6fdca545245e0cfca7897d1c21e7425dd63d896