Correct confusing naming in _calculate_cw_shard_sizes_and_offsets (#965)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/965

`hash_size` are now used in both `_calculate_rw_shard_sizes_and_offsets` and `_calculate_cw_shard_sizes_and_offsets`, but carry different meanings. It means `rows` in the former (this is correct) but means `columns` in the latter (this is wrong). We should make it correct and consistent.

Reviewed By: joshuadeng

Differential Revision: D42640749

fbshipit-source-id: 0248f2fca927694e0cdccab1c6e6a10d966a1db9