Test for sharded QEBC to be fxtrace-jitscript compatible (#960)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/960

Base test to make sharded QEBC to be fx traceable and jit scriptable.

Current model for testing is TestSparseNN from test_models wrapped in another module that has only Tensor-bound arguments to skip dataclass ModelInput.

FX-JIT workarounds:

**1. dataclasses** are not torch jit scriptable as they contain generated __init__ without source - jit will not have access for it and fail

For future we can think about workaround there or refactor out dataclasses in torchrec if they become blocker for scripting

**2. DMP has  varargs forward(args,kwargs)**
Using `exec` to define forward with the same signature (and jit annotations) as dmp_wrapped_module.

With this code state DMP can be fx-jit only if it is a root module. In future we can extend this solution for DMPs not on root level.

**3. Iterating through proxies**
Multiple places converted manualy to iterate `for i in range(len(...))` style, adding torch.fx.wrap("len") on the module level.

**4. torch jit has only torch.device ctor from string**

Added workaround in fx.create_arg() to use string instead of torch.device object. Jit will explicitly convert that string to jit.torch.device

**5. LazyAwaitable**

Adding logic to call wait of LazyAwaitable when we create fx.create_arg.
This is not equivalent behaviour to eager mode when wait is called on __torch__function or the first access to getattr() of LazyAwaitable.
Logic in the diff can call wait() before it is called in eager.
Solution with equivalent logic is in progress.

**6 fx.path_of_module**

FX needs identical paths for each torch.nn.Module, after sharding torchrec adds modules which are not registered in named_modules().
Here we have a hack to use `f"_torchrec_fake_module_path_{mod.__class__}_{id(mod)}"` for unregistered modules.
Proper solution with fx_path and identical names is in progress

Reviewed By: s4ayub

Differential Revision:
D42192307

Privacy Context Container: L1138451

fbshipit-source-id: 086099ef9664bc8b2b225d599cfb6a1327c3fab2