Fix bug for weighted features in ETC
Summary: Fixes bug for ETC where some modules have weighted features and some do not. Current behavior is correct for if it is a PEA module (accepts both weighted and unweighted simultaneously), but if the ETC has a combination of weighted and unweighted modules, we don't assign the weighted feature to the weighted module.

Differential Revision: D42812654

fbshipit-source-id: 328d8606f463d674c32ac2aabd103a442a54fd75