Fix bug for weighted features in ETC (#989)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/989

Fixes bug for ETC where some modules have weighted features and some do not. Current behavior is correct for if it is a PEA module (accepts both weighted and unweighted simultaneously), but if the ETC has a combination of weighted and unweighted modules, we don't assign the weighted feature to the weighted module.

Reviewed By: colin2328

Differential Revision: D42812654

fbshipit-source-id: 937de28fea12190dc63c17377bf646cfad42527a