Even out UVM/CPU memory pressure on different hosts (#1012)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1012

Some model is using UVM extensively. The problem with UVM + trec sharder today is that it's allocating UVM tables in order, and doesn't consider the host-level CPU memory pressure. In addition, the ddr limit is per rank. This creates a problem:

1. The big UVM tables will be assigned to ranks in order, so rank 0 will get overloaded and end up OOMing (due to in-trainer publish, async cp, etc)
2. We cannot lower the per-rank limit (128GB now), because it will limit the biggest UVM table we can put - a rank can take a 400GB table just fine (just that the other ranks should not take more UVM tables). if we lower it, video job will fail.

As a workaround, we'll just use the rank % local_world_size as a secondary key. This way the table will be allocated in the order of

host0_rank0 -> host1_rank0 -> host0_rank1 -> host1_rank1

This way we can even out the pressure on each host

Reviewed By: jianyuh, ge0405

Differential Revision: D43024113

fbshipit-source-id: 3ccf94de5e0334cafadcf6a9ff7947f7ee1193d1