Fix planning for sequence embeddings in towers (#1032)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1032

Currently when planning we do label sequence embeddings as pooled if they are within an EmbeddingTower or EmbeddingTowerCollection.

When checking if a ShardingOption is pooled and the module that contains the parameter is an ET or ETC when we loop through the named modules we check to see if the parameter that is the same name as the module and if so we check if it is an EmbeddingCollection to determine if its a sequence embedding. However, the name belongs to the parameter which is a child of the EmbeddingCollection module so the names will never match.
i.e.
{F869817109}

Reviewed By: dstaay-fb

Differential Revision: D43202613

fbshipit-source-id: a0e434deaab053a72ec6278080182507801922d6