Fix shard rank mismatch issue in per table optimizer for ETC
Summary:
We encountered this error:
`Number of local shards (1) does not match number of local shards metadata in sharded_tensor_metadata (0) on rank (8)`
in the per table EmbeddingFusedOptimizer path when calling `ShardedTensor._init_from_local_shards_and_global_metadata` for an ETC model.

This issue occurred for ETC because for parameters inside ETC the local shard metadata and global metadata are within the local world size, so they only exist on a world of size 8 with ranks 0-7. This is problematic for non host_0 hosts because the rank from the pg would correspond to the overall world size and when trying to match the local metadata with global metadata the global rank wouldn't exist for any shard in the global metadata.

However, we resolve the issue by passing in the correct intra-host pg from ETC, which along the sharded metadata only knows the local world (size 8 and ranks 0-7). Since the current code did not pass in a pg it used the default pg which had the global view.

Differential Revision: D43210984

fbshipit-source-id: 0b20c0d24486bc3fcc499015cdb6d7e90dda2017