Refactor tests to reduce duplicate code (#1043)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1043

Notice that our tests have shared the same logic to compute and test the metric values. I refactor it into a common utility function `metric_test_helper` and uses it for various metric tests. The result is we are able to delete 776 lines. Moreover, in the future, we no longer have to re-write pretty much the same code for new metric tests.

Reviewed By: fegin, joshuadeng, yachyv7

Differential Revision:
D43381828

Privacy Context Container: L1078771

fbshipit-source-id: a82403a4b9445a234d1a5896ec8110da2c287482