Add logloss metric (#1054)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1054

This diff implements the logloss metrics in RecMetrics's NE. User can toggle the `include_logloss` flag to get logloss metrics, on top of NE.
# Logloss
The logloss metric is the log of cross entropy loss. It can be represented mathematically as follow:
{F869642849}
# Implementation
The implementation in this diff is exactly the same as Caffe2 implementation. It differs the mathematical representation in that it divides the above term by the number of examples:
```
logloss / (numPositiveLabels + numNegativeLabels)
```
Caffe2 code pointer: https://fburl.com/code/p7udvje3

Reference doc: https://fb.quip.com/mLLRA037sbFd

Reviewed By: colin2328

Differential Revision: D43136575

fbshipit-source-id: aaa3265694f9dfcc648649e495b246b13dede7e6