fix test_sharding_fused_ebc (#1066)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1066

`test_sharding_fused_ebc` would fail because tensors are registered in List instead of ModuleList. And thus copying it over to it's respective device would cause the test to fail. We fix this by initializing and keeping `fused_ebc` on CPU and moving it over to respective cuda rank only in the DMP call when sharding. We also only then move the `kjt_input` over to GPU when obtaining preds from the sharded model

Reviewed By: joshuadeng

Differential Revision: D43780811

fbshipit-source-id: 94d1fb65e92b900de7388a1e642596895eb57756