Set memcpy and data dist streams to high priority in train pipeline (#1076)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1076

In some instances kernels in memcpy and data dist streams will not be scheduled immediately and can block the CPU. The kernels scheduled on these streams are few and lightweight, so it should not affect the main stream to have them set as high priority.

Example:
{F902622279}

KJTListSplitsAwaitable will be blocked on cpu waiting for a sync to be done in the data dist stream. However, nothing is blocking the sync from being done in parallel, it is only due to priority that it is getting scheduled later waiting redundantly on other streams to finish.

Reviewed By: bigning

Differential Revision: D44009082

fbshipit-source-id: 5693b7aa5652be965d6535330b3e6b23a7c2a8b1