Add comm data type size for both forward and backward comms (#1093)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1093

Current planner comms cost estimation does not take into the account of communication data type. This means that the cost estimation is not accurate when the communication data type is different from the table storage data type (e.g. fp8 communication).

We add the following variables for each ShardingOption in planner's perf functions:
* table_data_type_size: the data type size of the embedding table (e.g. float16 -> 2 bytes)
* fwd_a2a_comm_data_type_size, bwd_a2a_comm_data_type_sizem fwd_sr_comm_data_type_size, bwd_sr_comm_data_type_size: they are obtained from the quantized comms codec registry depending on the sharding type and if they are pooled.

Note that
* We did not change the storage estimates functions
* We did not estimate the encoding and decoding cost

Reviewed By: ge0405

Differential Revision: D44227825

fbshipit-source-id: 8f5c060b4cf81c512742a48c067a96fbbe50e05f