Start input dist before fwd in train pipeline and prevent unnecessary wait for batch (#1112)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1112

In order to improve perf when working on optimizing the FX transform free model, we had to efficiently handle 13 KJTs in the input dist and minimize input dist latency by avoiding CPU blocked time.

In the current TrainPipeline implementation the input dist occurs after the forward pass but before the backward. The input dist consisting of the splits a2a, the subsequent wait(), and then issuing the tensors a2as. The wait()s for the tensors all to all get called in the following forward call.

{F943438400}

From the figure above we see that wait() for the splits all to all blocks the CPU for a while as it is waiting for the splits a2a collective to finish. To avoid this CPU blockage, we split up the input dist by issuing the splits a2a before the forward and then keeping the wait() for the splits a2a and issuing the tensors a2as between the forward and backwards.

This optimization allows the splits a2a collective to be scheduled earlier and finish by the time the rest of the input dist continues, so that the CPU does not have to be blocked waiting for the splits a2a collective to finish.

However, this introduced a slightly different issue:
{F943478391}

After moving the start of the input dist, issuing the splits a2a, before the forwards we ran into another CPU blockage. This time it was waiting for the copy_batch_to_gpu on the memcpy stream to finish its device transfer for batch_i+2. However, this was actually a redundant wait, since the wait_for_batch before the input dist was there to ensure the device transfer had finished for the input dist batch_i+1 on the memcpy stream. This was actually a bug because since the copy_batch_to_gpu started before input dist, it would be using memcpy stream on batch_i+2 and input dist would needlessly call wait_for_batch on memcpy stream for batch_i+2.

To mitigate this bug we move the start of input dist before copy_batch_to_gpu, ensuring the wait_for_batch inside input dist would be waiting on the correct batch, and not blocking incorrectly. After this we finally reach a clean state where the input dist gpu work becomes hidden.
{F943436776}

Reviewed By: bigning

Differential Revision: D44781986

fbshipit-source-id: 2741d002eafde2591681908da9eacdae97bf39fe