Quant modules FX syntax to preserve TBE in GraphModule (#1129)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1129

The change is similar to D44961148 but for non-sharded quant EBC/EC.

We want to keep TBEs unchanged in fx.GraphModule to have access to all their fields.
For that we must call it as ebc(...) to generate call_module, but that requires module to be registered => check that self.tbe_registered.

Reviewed By: s4ayub

Differential Revision:
D45118008

Privacy Context Container: L1138451

fbshipit-source-id: 70c196ee26b0afe6406cb58e35a5f2918c314f47