wrap dict update in PositionWeightedModule (#1113)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1113

The forward method of PositionWeightedModule does an in-place dict update. However, when this module is symbolically traced with torch.fx, it results in generated code like this:
```
dict = ...
update = dict.update({...})
user_of_dict = dict.get(...)
```
During FX graph manipulation, the `user_of_dict` node is attached to `dict.users`, instead of `update`. While this is valid at runtime, it presents a problem with various FX operations including `eliminate_dead_code()` and `legalize_graph()`, because `update` is impure but also has 0 users.

In the case of legalize_graph, this can result in a FX-valid runtime-invalid program due to the presence of side effects.

To prevent this, we can wrap `dict.update` in a method that is hidden from the FX tracer via `torch.fx.wrap`. This way, the proxy object that is passed in place of `features` var in the forward method will not be the same proxy object that is returned by the forward method. This will consequently change the codegenerated forward method to look something like:

```
dict = ...
update = wrapped_function_that_updates_dict({...})
user_of_dict = update.get(...)
```

Differential Revision: D44896165

fbshipit-source-id: 04f9ea4099f9b9c3c8110ec11712ab753fb0c22e