Keep an almost sorted list of devices (#1142)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1142

Currently, during the partitioning of a proposal, we shallow copy the list of devices and sort it for every shard in every sharding option. While we can't really avoid sorting this many times, we don't have to sort from scratch every time. Sorting an almost sorted list is supposed to be faster.

This diff keeps a list of almost sorted devices around so re-sorting takes less time.

This change should not affect the resulting plan.

Reviewed By: bigning, RenfeiChen-FB

Differential Revision: D45372819

fbshipit-source-id: 80f94d334fe3e84bd5e24661a8536a4a1ccb1917