Remove unncessary wait typing from quant modules (#1146)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1146

Currently structure for input_dist/output_dist for quant modules are a bit confusing because they try to adhere to the Awaitable structure of training modules. However, fundamentally they are not using collective calls and these don't make sense. We get around the type issue with NoWaits, which is confusing - so getting rid of them.

Reviewed By: IvanKobzarev

Differential Revision: D45449316

fbshipit-source-id: a39105b10a7bd0a2bd3ed5b17ea2b17b81797ad9