Fix optional typing for weights list in table-batched embedding op (#1159)
Summary:
X-link: https://github.com/pytorch/FBGEMM/pull/1762

Pull Request resolved: https://github.com/pytorch/torchrec/pull/1159

we were working with this api and found the type inconsistency a bit confusing at first. As the second tensor should only be relevant for quantized weights, we should be able to set this as optional in the case of using non-quantized embedding tables.

Reviewed By: q10, r-barnes, YLGH

Differential Revision: D45703587

fbshipit-source-id: 356eae325ad749a30a6e471e24fb0ed9ba9319a4