Revert D45006690: Multisect successfully blamed D45006690 for test or build failures (#1166)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1166

This diff is reverting D45006690
Depends on D45802234
D45006690: Zero allocation torchrec sharding via meta [QEBC] by s4ayub has been identified to be causing the following test or build failures:

Tests affected:
- [dper_lib/silvertorch/delayed_modules/wrapper_modules/tests:build_prospector_model_test - test_ig_prospector_with_atm_publish (dper_lib.silvertorch.delayed_modules.wrapper_modules.tests.build_prospector_model_test.TestBuildProspectorModel)](https://www.internalfb.com/intern/test/281475065803150/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2047271
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: pengtxiafb

Differential Revision: D45802243

fbshipit-source-id: 66c71c9c18c76d51cac7104a1c8b56cf6b13510d