Improve planner error print when the largest table cannot fit in a device at all or when storage reservation is too restrictive
Summary:
Add a warning when it is not possible to fit a table in the devices at all. E.g., when we have an extremely large table but we are doing table_wise sharding, or when storage reservation is too restrictive.

In that case, we print out the name of the table and the sharding_type. We also print out the percentage of storage reseration.

Also print the last planner error.

Note that this would still work when there are more than one proposal.

Reviewed By: ge0405

Differential Revision: D46285104

fbshipit-source-id: 263465d3457000394156e541a6fab3685a789a91