Revert D46566230: Multisect successfully blamed D46566230 for test or build failures (#1266)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1266

This diff is reverting D46566230
D46566230: [torchrec] Fix the has_feature_processor flag bug that makes every table weighted by henrylhtsang has been identified to be causing the following test or build failures:

Tests affected:
- [aps_models/ads/icvr/tests:icvr_fqn_test - test_fqn (aps_models.ads.icvr.tests.icvr_fqn_test.ICVRFqnTest)](https://www.internalfb.com/intern/test/281475073983925/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2365517
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D47144345

fbshipit-source-id: 0b324a1fe834dbedae981d7015cc4462ab2b107b