Find minimal col dim when its not provided (#1295)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1295

This diff has two parts.

In the first part, we handle the case when the per table base_dim is not given. We use a for loop for find the base_dim that is at least 128, divides the table dim, and is divisible by 4. This helps with the case when the table dim is not divisible by 128, since in that case the current code would just use the table dim.

In the second part, we replace `col_wise_shard_dim` with `_find_base_dim(col_wise_shard_dim, columns)`. In cases where `col_wise_shard_dim` is correctly provided, this does nothing. But this allows the user to put in a more flexible `col_wise_shard_dim`. For example, the users can just put in `col_wise_shard_dim = 40` for all tables, and the planner will find the per table base dim that (1) divides the table dim, (2) is divisible by 4 and (3) is larger than or equal to 40.

Reviewed By: bigning

Differential Revision: D47887146

fbshipit-source-id: deaafa143a03f14c9dcada17743033c623c8ae68