Add pytree support to KeyedJaggedTensor (#1287)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1287

This makes it possible to export models that take KeyedJaggedTensor
as input.

For folks who don't know pytrees: pytrees are an abstraction that allow you to easily run functions on all of the internal contents of a data structure. For example, a tree_map on a tuple would apply an operation to every element of the tuple. To make KJTs pytree'able, I simply register every field on pytree as something that pytree knows about.

Reviewed By: dstaay-fb

Differential Revision: D47769501

fbshipit-source-id: 30c17d4bc678047e255d413a4870b3434364c730