Add an unsync method to KeyedJaggedTensor
Summary:
In some situations, I need to clear the cached fields on a KeyedJaggedTensor.
In particular, if I'm passing a KJT to tracing, clearing the cached fields
ensures that I trace the recomputation of these fields inside the model,
which is typically what I want because otherwise I do not know that the fields
are related during model tracing.

Reviewed By: dstaay-fb

Differential Revision: D47797243

fbshipit-source-id: 361edfcb031323add88e16b719b74800c687459e