RW sharding compatible with meta sharding (#1315)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1315

The reason of the change:

tgif does "meta" sharding, specifying the device="meta".
But tensors of "meta" device are only handled for TBEs in runtime => as a result rw_sharding will register "meta" tensor without content and it will fail in runtime as it will not be replaced.

The solution:
Creating tensor for arguments only at first runtime using specified device (at runtime that will be cuda:{idx}) and cache it in dictionary that is held by the module.

torch.fx.wrap to make fx pass through the control logic of caching

Reviewed By: s4ayub

Differential Revision:
D47846180

Privacy Context Container: L1138451

fbshipit-source-id: 26b95a140ce854aef1461ac12f264e5ba4d9bb91