Add an unsync method to KeyedJaggedTensor (#1311)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1311

In some situations, I need to clear the cached fields on a KeyedJaggedTensor.
In particular, if I'm passing a KJT to tracing, clearing the cached fields
ensures that I trace the recomputation of these fields inside the model,
which is typically what I want because otherwise I do not know that the fields
are related during model tracing.

Reviewed By: dstaay-fb

Differential Revision: D47797243

fbshipit-source-id: a9b54e6d11ba2f2895dee157f9468076faac73f2