MemoryBalancedPartitioner (#1302)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1302

Refactored Dharak's diffs D47073930 and D47567911.

Create a partitioner that calls the default GreedyPerfPartitioner a few times with different storage constraint, using binary search, to figure out a sharding plan that uses the least memory provided that the perf isn't impacted a lot.

Note that here `reserved_hbm_size` still matters, unlike the original diff. We recommend using a small `reserved_hbm_size` and allow the partitioner to figure out how much hbm storage is actually needed.

This diff only add the partitioner to the library. It does not enable it.

Reviewed By: bigning, dharakk

Differential Revision: D47731441

fbshipit-source-id: b2d6002e4934d1a2c05268cae1c7b5f8704f9e80