Add the GRADIENT_DIVISION for sequence embedding a2a (#1322)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1322

credits to YLGH who found that we are missing the gardient division for sequence embedding. We also investigated why the existing seq embedding unit tests didn't catch it. We found it was because the sequence embedding contributed very little to the final prediction and loss, so although the sharded model seqeunce embedding gradient don't match the unsharded model sequence embedding, but after the optimizer step, the final prediction is still the same. To fix this, we simply multiply 10 to the sequence embedding output.

Reviewed By: colin2328

Differential Revision: D48090579

fbshipit-source-id: 0a5ee4e7620c5a79d67ab0d8fd15e52b1e4b3327