Update KJT treespec serialization (#1354)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1354

We hardened the treespec serialization system + registration with
https://github.com/pytorch/pytorch/pull/106116. Will update the KJT treespec
serialization now.

OSS update is that a default `json.dumps` serialization will be used to serialize the
context, and the treespec type will be automatically serialized using
`f"{type.__module}.{type.__name__}"`. Users who want to customize the
serialization, or who do not have a json-dumpable context, can pass in a
`to_dumpable_context` function to convert the context to a json dumpable
format.

Since the KJT serialization is already using `json.dumps`, this implies the
context is already json serializable, so we can just use the default
serialization here.

Reviewed By: zhxchen17

Differential Revision: D48747343

fbshipit-source-id: dc6a67d3860014e28d5214235413c128a98d50ab